Remove .semgrepignore support
.semgrepignore was a stop gap to get file ignores while
semgrep did not yet have excludes/includes properly implemented

Now that semgrep has this functionality it is fine to remove here
so we don't have to support it in different places. Also the
include/exclude in semgrep is much more efficient since it does
not need to walk the entire file tree.