Add finding index to findings hash
In order to not trigger a new finding on whitespace changes and line
moves, we exclude surrounding context data from findings hashes.
However, this means that duplicate results with the exact same code in
a file will show up as only one finding, which is not desirable.

Here, we include the index of a finding within the file, so these show
up as separate findings.

This was tested by committing a "bad.py" with:
```
  5 == 5
```
Then committing an edit to that file to read:
```
  5 == 5

  5 == 5
```
We validated that the agent found 2 findings in the second commit, and
1 new finding when comparing the second commit to the first.

Next, I added a comment between the two lines in the file, and validated
that no new findings were discoverd.

I then validated the finding number sent to the backend with:
```
    semgrep-agent \
      --config r/python.lang.correctness.useless-eqeq.useless-eqeq \
      --baseline-ref head^^ \
      --publish-url "https://dev.semgrep.dev" \
      --publish-token '********' \
      --publish-deployment 1
```