Calculate merge base in GHA
Fixes an issue where we were baselining against the merge head, instead
of the merge base.

This would cause us to report issues that appear in the merge target
branch in addition to issues on the PR.

Fix by baselining to the merge base as reported by `git merge-base`.

Tested on https://github.com/returntocorp/semgrep-app/pull/728.