Allow override of "commit" reported to the backend
This commit is used to generate linking URLs, and users may want to
override the reported commit.

For example, when running on Buildkite, a user should run the following
to find the correct issues in a pull request:

  git merge base
  semgrep-agent --baseline-ref base

However, this will report the merge commit to the backend, whereas we
should instead report the commit SHA _before_ merge.

To override, a user would now set the SEMGREP_COMMIT environment
variable:

  SEMGREP_COMMIT=my_sha semgrep-agent --baseline-ref base