Our output has confused many users in the past, because pre-existing issues are discussed in the output but never uploaded to the app. This PR is a first step toward clarifying what we are doing with the diffing logic, so that not all users have to comb our docs to figure it out.
Long-term there are huge improvements we can still make here by
showing total number of issues in one place, and breakdowns by
- blocking/non-blocking
- muted/not-muted
- pre-existing/new

The way we print things out right now follows the order in which
our code calculates things, which is not the most intuitive way
to summarize this information. I downscoped making this very good,
since I know the goal is to move a lot of this into semgrep itself.